Replace std::cout with Core::Logging for more control over output
In this commit, we replaced the usage of std::cout and std::cerr with Core::Logging in several places to digest gateway messages and Websocket OpCodes. This allows for more granular control over the output, as well as providing a consistent framework for debugging and logging. These changes will greatly enhance our ability to track, report, and handle potential issues in the future.